requirements.txt
Django==5.0.6
djangorestframework==3.15.2

psycopg2-binary==2.9.9
gunicorn==21.2.0

celery==5.4.0
redis==5.0.6

Faker==25.0.1

pytest==8.3.2
pytest-django==4.8.0
